+++ README.md
@@ -24,11 +24,10 @@ Compiled and production-ready code can be found in the `dist` directory. The `sr
 
 ```html
 <script src="dist/js/classList.js"></script>
-<script src="dist/js/buoy.js"></script>
 <script src="dist/js/gumshoe.js"></script>
 ```
 
-Gumshoe requires [classList.js](https://github.com/eligrey/classList.js), a polyfill that extends ECMAScript 5 API support to more browsers. It also requires [Buoy](https://github.com/cferdinandi/buoy), a lightweight collection of helper methods for getting stuff done with native JavaScript.
+Gumshoe requires [classList.js](https://github.com/eligrey/classList.js), a polyfill that extends ECMAScript 5 API support to more browsers.
 
 ### 2. Add the markup to your HTML.
