+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "gumshoe",
-	"version": "1.2.0",
+	"version": "2.0.0",
 	"description": "A simple, framework-agnostic scrollspy script.",
 	"main": "./dist",
 	"author": {
