+++ test/results/unit-tests.html
@@ -12,7 +12,7 @@
         <td colspan="3" title="Mozilla/5.0 (Macintosh; Intel Mac OS X) AppleWebKit/534.34 (KHTML, like Gecko) PhantomJS/1.9.8 Safari/534.34">Browser: PhantomJS 1.9.8 (Mac OS X 0.0.0)</td>
       </tr>
       <tr class="overview">
-        <td colspan="3">Timestamp: 7/2/2015, 8:21:40 PM</td>
+        <td colspan="3">Timestamp: 7/5/2015, 8:56:10 PM</td>
       </tr>
       <tr>
         <td colspan="3">
@@ -20,7 +20,7 @@
           0 errors / 
           0 failures / 
           0 skipped / 
-          runtime: 0.002s
+          runtime: 0.003s
         </td>
       </tr>
       <tr class="header">
@@ -29,7 +29,7 @@
         <td>Suite / Results</td>
       </tr>
       <tr class="pass">
-        <td>Passed in 0.002s</td>
+        <td>Passed in 0.003s</td>
         <td>Document should include the gumshoe module</td>
         <td>gumshoe &raquo; Should initialize plugin</td>
       </tr>
